initial commit
this is still very rough and messy.  the typescript hasnt been separated
out into separate files yet and no bundle has been made.  I'd like to
wait until we get to a point that we need to split the module
definitions into their own files and then use AMD to make that
transition